+++ backend/tests/permissions.test.js
@@ -1,7 +1,11 @@
 const request = require('supertest')
 const mongoose = require('mongoose')
 const jwt = require('jsonwebtoken')
-const { MongoMemoryServer } = require('mongodb-memory-server')
+const {
+  startMongoMemoryServer,
+  isMongooseConnected,
+  disconnectMongooseIfConnected,
+} = require('./mongoMemoryTestServer')
 
 const createApp = require('../app')
 const User = require('../models/User')
@@ -34,7 +38,7 @@ beforeAll(async () => {
   process.env.AUTH_RATE_LIMIT_MAX = '100000'
   process.env.DEFAULT_TENANT = TEST_TENANT
 
-  mongo = await MongoMemoryServer.create()
+  mongo = await startMongoMemoryServer()
   const mongoUri = mongo.getUri()
   process.env.MONGO_URI = mongoUri
   process.env.MONGO_URI_LOOPC = mongoUri
@@ -43,6 +47,7 @@ beforeAll(async () => {
 })
 
 afterEach(async () => {
+  if (!isMongooseConnected(mongoose)) return
   await Promise.all([
     User.deleteMany({}),
     Task.deleteMany({}),
@@ -51,7 +56,7 @@ afterEach(async () => {
 })
 
 afterAll(async () => {
-  await mongoose.disconnect()
+  await disconnectMongooseIfConnected(mongoose)
   if (mongo) await mongo.stop()
 })
